aggiornamento 2025 06 23 16:54 [vBETA 0.5.118]

+++ database/migrations/2025_04_06_165013_create_compositions_table.php
@@ -14,15 +14,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('compositions', function (Blueprint $table) {
-            $table->string('espansione', 10);
-            $table->integer('numero');
-            $table->integer('idMazzo');
-            $table->foreign('idMazzo')->references('id')->on('decks')->onDelete('cascade');
-            $table->boolean('foil')->default(0);
-            $table->integer('copie')->default(1);
-            $table->string('id', 20)->primary();
-        });
+        if (!Schema::hasTable('compositions')) {
+            Schema::create('compositions', function (Blueprint $table) {
+                $table->string('espansione', 10);
+                $table->integer('numero');
+                $table->integer('idMazzo');
+                $table->foreign('idMazzo')->references('id')->on('decks')->onDelete('cascade');
+                $table->boolean('foil')->default(0);
+                $table->integer('copie')->default(1);
+                $table->string('id', 20)->primary();
+            });
+        }
     }
 
     /**
